RL-Learner/requirements.txt
# ============================================================================
# RL-Learner Python 依赖清单
# 安装方式：pip install -r requirements.txt
# ============================================================================

# --- 深度学习框架 ---
torch>=2.0.0                    # PPO 训练
onnx>=1.14.0                    # ONNX 模型操作

# --- 数值计算 ---
numpy>=1.24.0                   # 数值计算

# --- gRPC 通信 ---
grpcio>=1.56.0                  # gRPC 运行时
grpcio-tools>=1.56.0            # protoc Python 插件（生成 pb2 代码）
protobuf>=4.23.0                # Protobuf Python 绑定

# --- 配置与工具 ---
pyyaml>=6.0                     # YAML 配置文件解析
